org.rulez.magwas.zenta.controller.tests/src/org/rulez/magwas/zenta/controller/info/AndJunctionInfoProviderTest.java
package org.rulez.magwas.zenta.controller.info;

import static org.junit.Assert.*;

import org.junit.Before;
import org.junit.Test;
import org.rulez.magwas.zenta.model.IZentaPackage;

public class AndJunctionInfoProviderTest {

	private AndJunctionInfoProvider fixture;
	
	@Before
	public void setUp() {
		fixture = new AndJunctionInfoProvider();
	}

	@Test
	public void provider_is_correct() {
		assertEquals(fixture.providerFor(), IZentaPackage.eINSTANCE.getAndJunction());
	}
	
	@Test
	public void default_name_is_correct() {
		assertEquals(fixture.getDefaultName(), "And Junction");
	}

	@Test
	public void default_short_name_is_correct() {
		assertEquals(fixture.getDefaultShortName(), "And Junction");
	}
	@Test
	public void image_info_is_correct() {
		assertEquals(fixture.getImageInfo(),"img/zenta/junction-and-16.png");
	}

	@Test
	public void color_is_correct() {
		assertEquals(fixture.getDefaultColorString(),"#000000");
	}

}
